Mini-stats update with functionality and flexibility (#2428)
* Mini-stats - separated file into per-class files.

* fixed typo in comment

* implemented new features

* lint fix

* small improvements

* small changes (white text, black clear color, smaller gaps between graphs)

* text labels are slightly darker than numbers

* tracking basic frame timing and drawCall count even in release build

* spacing updates

* black background when graph is disabled

* size adjustement
